requirements.txt
python-dotenv
pandas
numpy
ipykernel
matplotlib
seaborn
fastapi
uvicorn
streamlit
requests
scikit-learn
tqdm
openai
groq
pydantic